+++ metadata.rb
@@ -1,11 +1,11 @@
-name              "supervisor"
+name              "z_supervisor"
 maintainer        "Noah Kantrowitz"
 maintainer_email  "noah@opscode.com"
 license           "Apache 2.0"
 description       "Installs supervisor and provides resources to configure services"
-version           "0.4.12"
+version           "1.0.0"
 
-recipe "supervisor", "Installs and configures supervisord"
+recipe "z_supervisor", "Installs and configures supervisord"
 
 depends "python"
 depends "locale"